refactored itests names
