+++ modules/engines/engine-hdb/pom.xml
@@ -251,52 +251,13 @@
             <artifactId>maven-failsafe-plugin</artifactId>
             <version>3.0.0-M5</version>
             <configuration>
-              <includes>
-                <include>**/*ITCase</include>
-              </includes>
-            </configuration>
-            <executions>
-              <execution>
-                <id>integration-test-embedded</id>
-                <phase>integration-test</phase>
-                <goals>
-                  <goal>integration-test</goal>
-                </goals>
-                <inherited>false</inherited>
-              </execution>
-              <execution>
-                <id>integration-test-verify</id>
-                <phase>post-integration-test</phase>
-                <goals>
-                  <goal>verify</goal>
-                </goals>
-                <inherited>false</inherited>
-              </execution>
-            </executions>
-          </plugin>
-        </plugins>
-      </build>
-    </profile>
-
-    <profile>
-      <id>hana-itests</id>
-      <activation>
-        <activeByDefault>false</activeByDefault>
-      </activation>
-      <build>
-        <plugins>
-          <plugin>
-            <groupId>org.apache.maven.plugins</groupId>
-            <artifactId>maven-failsafe-plugin</artifactId>
-            <version>3.0.0-M5</version>
-            <configuration>
               <systemPropertyVariables>
                 <hana.url>url</hana.url>
                 <hana.username>username</hana.username>
                 <hana.password>password</hana.password>
               </systemPropertyVariables>
               <includes>
-                <include>**/*HanaCase</include>
+                <include>**/*ITTest</include>
               </includes>
             </configuration>
             <executions>
