+++ modules/engines/engine-hdb/src/test/java/com/sap/xsk/hdb/ds/test/itest/hdbsequence/XSKHDBSequenceParserPostgreSQLITTest.java
@@ -9,7 +9,7 @@
  * SPDX-FileCopyrightText: 2019-2021 SAP SE or an SAP affiliate company and XSK contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.sap.xsk.hdb.ds.test.itest.hdbview;
+package com.sap.xsk.hdb.ds.test.itest.hdbsequence;
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
@@ -21,30 +21,40 @@ import com.sap.xsk.hdb.ds.test.itest.model.JDBCModel;
 import com.sap.xsk.hdb.ds.test.itest.module.XSKHDBTestModule;
 import org.eclipse.dirigible.core.scheduler.api.SynchronizationException;
 import org.eclipse.dirigible.repository.local.LocalResource;
+import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import org.testcontainers.containers.MySQLContainer;
+import org.testcontainers.containers.Network;
+import org.testcontainers.containers.PostgreSQLContainer;
 import javax.sql.DataSource;
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.ArrayList;
+import java.util.List;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertFalse;
 
-public class XSKHDBViewParserMySQLITCase {
+/**
+ * XSKHDBSequenceParserITCase will test only sequence
+ * creation due existing sequence drop logic that is never invoked.
+ * Existing sequence alter logic is not checked due db limitations on retrieving sequence info.
+ */
+public class XSKHDBSequenceParserPostgreSQLITTest {
 
-  private static MySQLContainer jdbcContainer;
+  private static PostgreSQLContainer jdbcContainer;
   private static Connection connection;
   private static IXSKHDBCoreFacade facade;
 
 
   @BeforeClass
   public static void setUp() throws SQLException {
+    Network network = Network.newNetwork();
     jdbcContainer =
-        new MySQLContainer<>("mysql:5.7");
+        new PostgreSQLContainer<>("postgres:alpine");
     jdbcContainer.start();
     JDBCModel model = new JDBCModel(jdbcContainer.getDriverClassName(), jdbcContainer.getJdbcUrl(), jdbcContainer.getUsername(),
         jdbcContainer.getPassword());
@@ -53,24 +63,33 @@ public class XSKHDBViewParserMySQLITCase {
     facade = injector.getInstance(Key.get(IXSKHDBCoreFacade.class, Names.named("xskHDBCoreFacade")));
   }
 
-  @Test
-  public void testHDBViewCreate() throws XSKDataStructuresException, SynchronizationException, IOException, SQLException {
-    Statement stmt = connection.createStatement();
-    stmt.executeUpdate("create table `test`.`acme.com.test.tables::MY_TABLE1`(Column1 integer,Column2 integer)");
-    stmt.executeUpdate("create table `test`.`acme.com.test.views::MY_VIEW1`(Column1 integer,Column2 integer)");
+  @AfterClass
+  public static void cleanUp() {
+    jdbcContainer.stop();
+  }
 
+  @Test
+  public void testHDBSequenceCreate() throws XSKDataStructuresException, SynchronizationException, IOException, SQLException {
     LocalResource resource = XSKHDBTestModule.getResources("/usr/local/target/dirigible/repository/root",
-        "/registry/public/hdbview-itest/SampleMySQLXSClassicView.hdbview",
-        "/hdbview-itest/SampleMySQLXSClassicView.hdbview");
+        "/registry/public/sequence-itest/SampleSequence_HanaXSClassic.hdbsequence",
+        "/sequence-itest/SampleSequence_HanaXSClassic.hdbsequence");
 
     this.facade.handleResourceSynchronization(resource);
     this.facade.updateEntities();
 
-    ResultSet rs = stmt.executeQuery(String.format("SELECT COUNT(*) as rawsCount FROM %s", "`hdbview-itest::SampleMySQLXSClassicView`"));
-    assertTrue(rs.next());
-    assertEquals(0, rs.getInt("rawsCount"));
-    stmt.executeUpdate(String.format("DROP VIEW %s", "`hdbview-itest::SampleMySQLXSClassicView`"));
-    stmt.executeUpdate("drop table `test`.`acme.com.test.tables::MY_TABLE1`");
-    stmt.executeUpdate("drop table `test`.`acme.com.test.views::MY_VIEW1`");
+    Statement stmt = connection.createStatement();
+    List<String> dbSequences = new ArrayList<>();
+    ResultSet rs = stmt.executeQuery("SELECT  relname sequence_name FROM  pg_class WHERE  relkind = 'S'");
+    while (rs.next()) {
+      dbSequences.add(rs.getString("sequence_name"));
+    }
+    assertEquals(1, dbSequences.size());
+    assertEquals("sequence-itest::SampleSequence_HanaXSClassic", dbSequences.get(0));
+
+    stmt.executeUpdate(String.format("DROP SEQUENCE \"%s\"", dbSequences.get(0)));
+    rs = stmt.executeQuery("SELECT  relname sequence_name FROM  pg_class WHERE  relkind = 'S'");
+    assertFalse(rs.next());
   }
+
+
 }
