+++ modules/engines/engine-hdb/src/test/java/com/sap/xsk/hdb/ds/test/itest/hdbview/XSKHDBViewParserMySQLITTest.java
@@ -19,12 +19,11 @@ import com.sap.xsk.hdb.ds.api.XSKDataStructuresException;
 import com.sap.xsk.hdb.ds.facade.IXSKHDBCoreFacade;
 import com.sap.xsk.hdb.ds.test.itest.model.JDBCModel;
 import com.sap.xsk.hdb.ds.test.itest.module.XSKHDBTestModule;
-import com.sap.xsk.utils.XSKHDBUtils;
 import org.eclipse.dirigible.core.scheduler.api.SynchronizationException;
 import org.eclipse.dirigible.repository.local.LocalResource;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.containers.MySQLContainer;
 import javax.sql.DataSource;
 import java.io.IOException;
 import java.sql.Connection;
@@ -35,9 +34,9 @@ import java.sql.Statement;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-public class XSKHDBViewParserPostgreSQLITCase {
+public class XSKHDBViewParserMySQLITTest {
 
-  private static PostgreSQLContainer jdbcContainer;
+  private static MySQLContainer jdbcContainer;
   private static Connection connection;
   private static IXSKHDBCoreFacade facade;
 
@@ -45,7 +44,7 @@ public class XSKHDBViewParserPostgreSQLITCase {
   @BeforeClass
   public static void setUp() throws SQLException {
     jdbcContainer =
-        new PostgreSQLContainer<>("postgres:alpine");
+        new MySQLContainer<>("mysql:5.7");
     jdbcContainer.start();
     JDBCModel model = new JDBCModel(jdbcContainer.getDriverClassName(), jdbcContainer.getJdbcUrl(), jdbcContainer.getUsername(),
         jdbcContainer.getPassword());
@@ -54,26 +53,24 @@ public class XSKHDBViewParserPostgreSQLITCase {
     facade = injector.getInstance(Key.get(IXSKHDBCoreFacade.class, Names.named("xskHDBCoreFacade")));
   }
 
-
   @Test
   public void testHDBViewCreate() throws XSKDataStructuresException, SynchronizationException, IOException, SQLException {
     Statement stmt = connection.createStatement();
-    stmt.executeUpdate("create table \"public\".\"acme.com.test.tables::MY_TABLE1\"(Column1 integer,Column2 integer)");
-    stmt.executeUpdate("create table \"public\".\"acme.com.test.views::MY_VIEW1\"(Column1 integer,Column2 integer)");
+    stmt.executeUpdate("create table `test`.`acme.com.test.tables::MY_TABLE1`(Column1 integer,Column2 integer)");
+    stmt.executeUpdate("create table `test`.`acme.com.test.views::MY_VIEW1`(Column1 integer,Column2 integer)");
+
     LocalResource resource = XSKHDBTestModule.getResources("/usr/local/target/dirigible/repository/root",
-        "/registry/public/hdbview-itest/SamplePostgreXSClassicView.hdbview",
-        "/hdbview-itest/SamplePostgreXSClassicView.hdbview");
+        "/registry/public/hdbview-itest/SampleMySQLXSClassicView.hdbview",
+        "/hdbview-itest/SampleMySQLXSClassicView.hdbview");
 
     this.facade.handleResourceSynchronization(resource);
     this.facade.updateEntities();
 
-    ResultSet rs = stmt
-        .executeQuery(String.format("SELECT COUNT(*) as rawsCount FROM \"%s\"", "hdbview-itest::SamplePostgreXSClassicView"));
+    ResultSet rs = stmt.executeQuery(String.format("SELECT COUNT(*) as rawsCount FROM %s", "`hdbview-itest::SampleMySQLXSClassicView`"));
     assertTrue(rs.next());
     assertEquals(0, rs.getInt("rawsCount"));
-    stmt.executeUpdate(
-        String.format("DROP VIEW %s", XSKHDBUtils.escapeArtifactName(connection, "hdbview-itest::SamplePostgreXSClassicView")));
-    stmt.executeUpdate("drop table \"public\".\"acme.com.test.tables::MY_TABLE1\"");
-    stmt.executeUpdate("drop table \"public\".\"acme.com.test.views::MY_VIEW1\"");
+    stmt.executeUpdate(String.format("DROP VIEW %s", "`hdbview-itest::SampleMySQLXSClassicView`"));
+    stmt.executeUpdate("drop table `test`.`acme.com.test.tables::MY_TABLE1`");
+    stmt.executeUpdate("drop table `test`.`acme.com.test.views::MY_VIEW1`");
   }
 }
